backend/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
pydantic[email]==2.12.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
sqlalchemy==2.0.46
alembic==1.16.5
asyncpg==0.31.0
python-dotenv==1.2.1
aiohttp==3.13.3
bcrypt==5.0.0
